feat: Introduce bucket and metric aggregations (#336)
