+++ docs/aggregations/bucket.mdx
@@ -1,0 +1,162 @@
+---
+title: Bucket Aggregations
+---
+
+<Info>
+  **Prerequisite** Aggregations can only be performed over fields [indexed as
+  fast fields](/indexing/bm25).
+</Info>
+
+## Basic Usage
+
+```sql
+SELECT *
+FROM paradedb.aggregation('<index_name>', '<aggregation_query>');
+```
+
+<Accordion title="Example Usage">
+
+```sql
+SELECT *
+FROM paradedb.aggregation(
+    'idx_mock_items',
+    '{"aggs": {"histogram": {"field": "rating", "interval": 2}}}'
+)
+```
+
+</Accordion>
+
+<ParamField body="index_name">The name of the index.</ParamField>
+<ParamField body="aggregation_query">
+  The aggregation query string. Aggregation query strings follow the
+  ElasticSearch DSL.
+</ParamField>
+
+## Bucket Aggregation JSON Format
+
+### Histogram
+
+The `Histogram` aggregation returns a list of buckets, each representing a specific range of values in the
+aggregated field.
+
+```sql
+'{
+    "histogram": {
+        "field": "rating",
+        "interval": 2
+    }
+}'
+```
+
+<Expandable title="Config Options">
+
+<ParamField body="field" required type="string">
+  The name of the field to aggregate.
+</ParamField>
+<ParamField body="interval" required type="int">
+  The interval to chunk your data range. Must be a positive value.
+</ParamField>
+<ParamField body="offset" type="int">
+  Shifts the grid of buckets from `[interval * k, interval * (k + 1))` to
+  `[offset + interval * k, offset + interval * (k + 1))`. Must be in `[0,
+  interval)`. Example: For documents with values `9` and `12`, interval `10.0`,
+  they fall into buckets `0` and `10`. With offset `5`, both fall into bucket
+  `5`, range `[5..15)`.
+</ParamField>
+<ParamField body="min_doc_count" type="int">
+  Sets the minimum document count per bucket for inclusion in results. Defaults
+  to `0`.
+</ParamField>
+<ParamField body="hard_bounds" type="{'min': int, 'max': int}">
+  Constrains data to the `[min, max]` interval. Filters out values outside this
+  range. To enforce a specific range, set both `extended_bounds` and
+  `hard_bounds` to identical ranges.
+</ParamField>
+
+</Expandable>
+
+### Range
+
+The `Range` aggregation categorizes documents into specified value ranges, providing a count per range and
+enabling a clear, segmented view of data distribution across defined boundaries.
+
+```sql
+'{
+    "range": {
+        "field": "rating",
+        "ranges": [
+            { "to": 3.0 },
+            { "from": 3.0, "to": 7.0 }
+        ]
+    }
+}'
+```
+
+<Expandable title="Config Options">
+
+<ParamField body="field" required type="string">
+  The name of the field to aggregate.
+</ParamField>
+<ParamField body="ranges" required type="{'to': int, 'from': int}[]">
+  Note that this aggregation includes the from value and excludes the to value
+  for each range. Extra buckets will be created until the first to, and last
+  from, if necessary.
+</ParamField>
+<ParamField body="keyed" type="bool">
+  Whether to return the buckets as a hash map.
+</ParamField>
+
+</Expandable>
+
+### Terms
+
+The `Terms` aggregation creates a bucket for every unique term and counts the number of occurences.
+
+```sql
+'{
+    "terms": {
+        "field": "description"
+    }
+}'
+```
+
+<Expandable title="Config Options">
+
+<ParamField body="field" required type="string">
+  The field to aggregate on.
+</ParamField>
+<ParamField body="size" type="int">
+  The number of terms to return. Defaults to the top 10 terms with the most documents. Larger values for size are more expensive.
+</ParamField>
+<ParamField body="split_size" type="int">
+  Unused by tantivy.
+
+Since tantivy doesn’t know shards, this parameter is merely there to be used by consumers of tantivy. `shard_size` is the number of terms returned by each shard. The default value in Elasticsearch is `size * 1.5 + 10`.
+
+Should never be smaller than `size`.
+
+</ParamField>
+<ParamField body="segment_size" type="int">
+  To get more accurate results, we fetch more than `size` from each segment. Increasing this value will increase the cost for more accuracy. Defaults to `10 * size`.
+</ParamField>
+
+<ParamField body="show_term_doc_count_error" type="bool">
+  If you set the `show_term_doc_count_error` parameter to true, the terms
+  aggregation will include `doc_count_error_upper_bound`, which is an upper
+  bound to the error on the `doc_count` returned by each shard. It’s the sum of
+  the size of the largest bucket on each segment that didn’t fit into
+  `shard_size`. Defaults to true when ordering by count desc.
+</ParamField>
+
+<ParamField body="min_doc_count" type="int">
+  Filter out all terms with a count lower than `min_doc_count`. Defaults to 1.
+
+Note: Setting this to 0 will return all terms in the field, but it can be expensive.
+
+</ParamField>
+<ParamField body="missing" type="string">
+  Defines how documents that are missing a value should be treated. 
+  By default, they are ignored, but it is also possible to treat them as if they had a value.
+  For example, `"NO_DATA"`.
+</ParamField>
+</Expandable>
