+++ docs/aggregations/metrics.mdx
@@ -1,0 +1,168 @@
+---
+title: Metrics Aggregations
+---
+
+<Info>
+  **Prerequisite** Aggregations can only be performed over fields [indexed as
+  fast fields](/indexing/bm25).
+</Info>
+
+## Basic Usage
+
+```sql
+SELECT *
+FROM paradedb.aggregation('<index_name>', '<aggregation_query>');
+```
+
+<Accordion title="Example Usage">
+
+```sql
+SELECT *
+FROM paradedb.aggregation(
+    'idx_mock_items',
+    '{"aggs": {"avg": {"field": "rating"}}}'
+)
+```
+
+</Accordion>
+
+<ParamField body="index_name">The name of the index.</ParamField>
+<ParamField body="aggregation_query">
+  The aggregation query string. Aggregation query strings follow the
+  ElasticSearch DSL.
+</ParamField>
+
+## Metrics Aggregation JSON Format
+
+### Average
+
+```sql
+'{
+    "avg": {
+        "field": "rating"
+    }
+}'
+```
+
+<Expandable title="Config Options">
+  <ParamField body="field" required type="string">
+    The name of the field to aggregate.
+  </ParamField>
+  <ParamField body="missing" type="int">
+    The missing parameter defines how documents that are missing a value should
+    be treated. By default they will be ignored but it is also possible to treat
+    them as if they had a value.
+  </ParamField>
+</Expandable>
+
+### Count
+
+```sql
+'{
+    "value_count": {
+        "field": "rating"
+    }
+}'
+```
+
+<Expandable title="Config Options">
+  <ParamField body="field" required type="string">
+    The name of the field to aggregate.
+  </ParamField>
+  <ParamField body="missing" type="int">
+    The missing parameter defines how documents that are missing a value should
+    be treated. By default they will be ignored but it is also possible to treat
+    them as if they had a value.
+  </ParamField>
+</Expandable>
+
+### Max
+
+```sql
+'{
+    "max": {
+        "field": "rating"
+    }
+}'
+```
+
+<Expandable title="Config Options">
+  <ParamField body="field" required type="string">
+    The name of the field to aggregate.
+  </ParamField>
+  <ParamField body="missing" type="int">
+    The missing parameter defines how documents that are missing a value should
+    be treated. By default they will be ignored but it is also possible to treat
+    them as if they had a value.
+  </ParamField>
+</Expandable>
+
+### Min
+
+```sql
+'{
+    "min": {
+        "field": "rating"
+    }
+}'
+```
+
+<Expandable title="Config Options">
+  <ParamField body="field" required type="string">
+    The name of the field to aggregate.
+  </ParamField>
+  <ParamField body="missing" type="int">
+    The missing parameter defines how documents that are missing a value should
+    be treated. By default they will be ignored but it is also possible to treat
+    them as if they had a value.
+  </ParamField>
+</Expandable>
+
+### Stats
+
+```sql
+'{
+    "stats": {
+        "field": "rating"
+    }
+}'
+```
+
+<Expandable title="Config Options">
+  <ParamField body="field" required type="string">
+    The name of the field to aggregate.
+  </ParamField>
+  <ParamField body="missing" type="int">
+    The missing parameter defines how documents that are missing a value should
+    be treated. By default they will be ignored but it is also possible to treat
+    them as if they had a value.
+  </ParamField>
+</Expandable>
+
+### Percentiles
+
+```sql
+'{
+    "percentiles": {
+        "field": "rating"
+    }
+}'
+```
+
+<Expandable title="Config Options">
+  <ParamField body="field" required type="string">
+    The name of the field to aggregate.
+  </ParamField>
+  <ParamField body="percents" type="int[]">
+    The percentiles to compute. Defaults to `[1.0, 5.0, 25.0, 50.0, 75.0, 95.0,
+    99.0]`.
+  </ParamField>
+  <ParamField body="keyed" type="bool">
+    Whether to return the percentiles as a hash map.
+  </ParamField>
+  <ParamField body="missing" type="int">
+    The missing parameter defines how documents that are missing a value should
+    be treated. By default they will be ignored but it is also possible to treat
+    them as if they had a value.
+  </ParamField>
+</Expandable>
