+++ docs/aggregations/overview.mdx
@@ -1,0 +1,11 @@
+---
+title: Aggregations Explained
+---
+
+Aggregations are a way to perform analytics on your SQL data. While Postgres provides aggregations, these
+operations can take upwards of several minutes. ParadeDB leverage's Tantivy fast fields to perform
+lightning-fast aggregations on individual columns.
+
+There are two types of aggregations: bucket and metrics. Bucket aggregations are useful for categorizing
+and grouping data. Supported bucket aggregations are `Histogram`, `Range`, and `Terms`.
+Supported metrics aggregations are `Average`, `Count`, `Max`, `Min`, `Stats`, `Sum`, and `Percentiles`.
