+++ pg_bm25/src/api/index.rs
@@ -1,0 +1,84 @@
+use pgrx::{iter::TableIterator, *};
+use tantivy::schema::*;
+
+use crate::index_access::utils::get_parade_index;
+use crate::parade_index::fields::ToString;
+
+#[allow(clippy::type_complexity)]
+#[pg_extern]
+pub fn schema_bm25(
+    index_name: &str,
+) -> TableIterator<(
+    name!(name, String),
+    name!(field_type, String),
+    name!(stored, bool),
+    name!(indexed, bool),
+    name!(fast, bool),
+    name!(fieldnorms, bool),
+    name!(expand_dots, Option<bool>),
+    name!(tokenizer, Option<String>),
+    name!(record, Option<String>),
+    name!(normalizer, Option<String>),
+)> {
+    let parade_index = get_parade_index(index_name.to_string());
+    let underlying_index = parade_index.underlying_index;
+    let schema = underlying_index.schema();
+
+    let mut field_rows = Vec::new();
+
+    for field in schema.fields() {
+        let (field, field_entry) = field;
+        let name = schema.get_field_name(field).to_string();
+
+        let (field_type, tokenizer, record, normalizer, expand_dots) =
+            match field_entry.field_type() {
+                FieldType::I64(_) => ("I64".to_string(), None, None, None, None),
+                FieldType::U64(_) => ("U64".to_string(), None, None, None, None),
+                FieldType::F64(_) => ("F64".to_string(), None, None, None, None),
+                FieldType::Bool(_) => ("Bool".to_string(), None, None, None, None),
+                FieldType::Str(text_options) => {
+                    let indexing_options = text_options.get_indexing_options();
+                    let tokenizer = indexing_options.map(|opt| opt.tokenizer().to_string());
+                    let record = indexing_options.map(|opt| opt.index_option().to_string());
+                    let normalizer = text_options
+                        .get_fast_field_tokenizer_name()
+                        .map(|s| s.to_string());
+                    ("Str".to_string(), tokenizer, record, normalizer, None)
+                }
+                FieldType::JsonObject(json_options) => {
+                    let indexing_options = json_options.get_text_indexing_options();
+                    let tokenizer = indexing_options.map(|opt| opt.tokenizer().to_string());
+                    let record = indexing_options.map(|opt| opt.index_option().to_string());
+                    let normalizer = json_options
+                        .get_fast_field_tokenizer_name()
+                        .map(|s| s.to_string());
+                    let expand_dots = Some(json_options.is_expand_dots_enabled());
+                    (
+                        "JsonObject".to_string(),
+                        tokenizer,
+                        record,
+                        normalizer,
+                        expand_dots,
+                    )
+                }
+                _ => ("Other".to_string(), None, None, None, None),
+            };
+
+        let row = (
+            name,
+            field_type,
+            field_entry.is_stored(),
+            field_entry.is_indexed(),
+            field_entry.is_fast(),
+            field_entry.has_fieldnorms(),
+            expand_dots,
+            tokenizer,
+            record,
+            normalizer,
+        );
+
+        field_rows.push(row);
+    }
+
+    TableIterator::new(field_rows)
+}
