+++ pg_bm25/test/expected/aggregations.out
@@ -1,0 +1,14 @@
+-- Bucket aggregation 
+SELECT * FROM paradedb.aggregation('idxaggregations', '{"aggs": {"histogram": {"field": "rating", "interval": 2}}}');
+                                                     aggregation                                                     
+---------------------------------------------------------------------------------------------------------------------
+ {"aggs": {"buckets": [{"key": 0.0, "doc_count": 1}, {"key": 2.0, "doc_count": 12}, {"key": 4.0, "doc_count": 28}]}}
+(1 row)
+
+-- Metrics aggregation
+SELECT * FROM paradedb.aggregation('idxaggregations', '{"aggs": {"avg": {"field": "rating"}}}');
+               aggregation               
+-----------------------------------------
+ {"aggs": {"value": 3.8536585365853657}}
+(1 row)
+
