+++ pg_bm25/test/sql/aggregations.sql
@@ -1,0 +1,5 @@
+-- Bucket aggregation 
+SELECT * FROM paradedb.aggregation('idxaggregations', '{"aggs": {"histogram": {"field": "rating", "interval": 2}}}');
+
+-- Metrics aggregation
+SELECT * FROM paradedb.aggregation('idxaggregations', '{"aggs": {"avg": {"field": "rating"}}}');
